Allow override to default value for TLS client auth when a CA certificate is specified.
The hard-coded value was tls.RequireAndVerifyClientCert, which required a valid cert for
all API paths.  We need the health-check API path (the ping operation) to be accessible
without a valid client cert, so that we can use AWS health checks for the server
containers.

So add a TLSClientAuthType property to the generated server struct.  If it is
set, use that value for httpsServer.TLSConfig.ClientAuth.  We will specify
tls.VerifyClientCertIfGiven.  If the TLS client requests the health-check API,
and does not specify a cert, we will use middleware to allow it to continue.
All other APIs will require a cert, and will get a 401 Unauthorized if a valid
cert is not specified.

This mirrors the approach taken by the WADLs.